feat: migrate to next/image with blur placeholders and responsive sizes
